Move common API types and functions into shared
