+++ crates/orderbook/src/api/get_orders_by_tx.rs
@@ -1,6 +1,7 @@
-use crate::{api::convert_json_response, orderbook::Orderbook};
+use crate::orderbook::Orderbook;
 use anyhow::Result;
 use ethcontract::H256;
+use shared::api::{convert_json_response, ApiReply};
 use std::{convert::Infallible, sync::Arc};
 use warp::{Filter, Rejection};
 
@@ -10,7 +11,7 @@ pub fn get_orders_by_tx_request() -> impl Filter<Extract = (H256,), Error = Reje
 
 pub fn get_orders_by_tx(
     orderbook: Arc<Orderbook>,
-) -> impl Filter<Extract = (super::ApiReply,), Error = Rejection> + Clone {
+) -> impl Filter<Extract = (ApiReply,), Error = Rejection> + Clone {
     get_orders_by_tx_request().and_then(move |hash: H256| {
         let orderbook = orderbook.clone();
         async move {
