+++ crates/orderbook/src/api/get_user_orders.rs
@@ -1,7 +1,8 @@
-use crate::{api::convert_json_response, orderbook::Orderbook};
+use crate::orderbook::Orderbook;
 use anyhow::Result;
 use primitive_types::H160;
 use serde::Deserialize;
+use shared::api::{convert_json_response, ApiReply};
 use std::{convert::Infallible, sync::Arc};
 use warp::{hyper::StatusCode, reply::with_status, Filter, Rejection};
 
@@ -19,7 +20,7 @@ fn request() -> impl Filter<Extract = (H160, Query), Error = Rejection> + Clone 
 
 pub fn get_user_orders(
     orderbook: Arc<Orderbook>,
-) -> impl Filter<Extract = (super::ApiReply,), Error = Rejection> + Clone {
+) -> impl Filter<Extract = (ApiReply,), Error = Rejection> + Clone {
     request().and_then(move |owner: H160, query: Query| {
         let orderbook = orderbook.clone();
         async move {
